# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.userPreferencesSetting:
    properties:
      audio_quality:
        type: string
      local_cache_max_size:
        type: integer
      play_mode:
        type: string
      playlist_view_mode:
        type: string
      theme_mode:
        type: string
      volume:
        type: number
    type: object
...
  models.ErrorResponse:
    properties:
      detail:
        description: 详细错误信息（可选）
        example: 详细错误信息
        type: string
      error:
        description: 错误信息
        example: 操作失败
        type: string
    type: object
...
  models.SuccessResponse:
    properties:
      message:
        description: 成功信息
        example: 操作成功
        type: string
    type: object
  models.ThemePackColors:
    properties:
      backgroundColor:
        type: string
      glassColor:
        description: |-
          GlassColor independently drives the Liquid Glass decorative tint
          (--glass-glow / --glass-glow-faint / --glass-sheen in the Lynx client),
          separate from SeedColor which drives the button/accent channel. Optional:
          when absent the client falls back to its star-blue glass baseline, so
          glass stays a different colour from buttons (true dual-channel). Existing
          packs without this field still validate.
        type: string
      seedColor:
        type: string
      surfaceColor:
        type: string
    type: object
  models.ThemePackData:
    properties:
      author:
        type: string
      cardRadius:
        type: number
      controlRadius:
        type: number
      dark:
        $ref: '#/definitions/models.ThemePackColors'
      description:
        type: string
      id:
        type: string
      light:
        $ref: '#/definitions/models.ThemePackColors'
      name:
        type: string
      navigationRadius:
        type: number
      navigationStyle:
        type: string
      playerGradient:
        items:
          type: string
        type: array
      schemaVersion:
        type: integer
      version:
        type: string
    type: object
...
  services.ThemePackListItem:
    properties:
      author:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      schema_version:
        type: integer
      theme_id:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
  services.ThemePackResponse:
    properties:
      author:
        type: string
      created_at:
        type: string
      data:
        $ref: '#/definitions/models.ThemePackData'
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      schema_version:
        type: integer
      theme_id:
        type: string
      updated_at:
        type: string
      version:
        type: string
    type: object
...
host: localhost:58091
info:
  contact:
    email: im.hanxi@gmail.com
    name: API Support
    url: https://github.com/songloft-org/songloft
  description: 轻量级音乐服务器 API 文档，支持本地音乐管理、网络歌曲、电台和歌单功能
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Songloft API
  version: 2.12.1
paths:
...
  /settings/user-preferences:
    get:
      description: 获取用户跨设备同步的偏好设置，包括主题、播放模式、视图模式、音质、缓存上限和音量。未配置时返回默认值。
      produces:
      - application/json
      responses:
        "200":
          description: 用户偏好设置
          schema:
            $ref: '#/definitions/handlers.userPreferencesSetting'
      security:
      - BearerAuth: []
      summary: 获取用户偏好设置
      tags:
      - 设置
    put:
      consumes:
      - application/json
      description: 保存用户跨设备同步的偏好设置。客户端登录后拉取、修改偏好时推送，实现多设备间偏好同步。
      parameters:
      - description: 用户偏好设置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.userPreferencesSetting'
      produces:
      - application/json
      responses:
        "200":
          description: 保存后的用户偏好设置
          schema:
            $ref: '#/definitions/handlers.userPreferencesSetting'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: 保存配置失败
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 保存用户偏好设置
      tags:
      - 设置
...
  /theme-packs:
    get:
      description: 返回所有已安装的主题包列表
      produces:
      - application/json
      responses:
        "200":
          description: 主题包列表
          schema:
            items:
              $ref: '#/definitions/services.ThemePackListItem'
            type: array
      security:
      - BearerAuth: []
      summary: 列出主题包
      tags:
      - 主题包
    post:
      consumes:
      - application/json
      description: 接收 .songloft-theme JSON 内容，验证后存入。相同 ID 的主题包会被更新。
      parameters:
      - description: 主题包 JSON（完整 .songloft-theme 文件内容）
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 导入成功
          schema:
            $ref: '#/definitions/services.ThemePackResponse'
        "400":
          description: 格式错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 导入主题包
      tags:
      - 主题包
  /theme-packs/{themeID}:
    delete:
      description: 删除指定主题包，若为当前激活主题则同时恢复默认
      parameters:
      - description: 主题包 ID
        in: path
        name: themeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "404":
          description: 主题包不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除主题包
      tags:
      - 主题包
    get:
      description: 返回主题包的完整信息，含主题配置数据
      parameters:
      - description: 主题包 ID
        in: path
        name: themeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 主题包详情
          schema:
            $ref: '#/definitions/services.ThemePackResponse'
        "404":
          description: 主题包不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取主题包详情
      tags:
      - 主题包
  /theme-packs/active:
    delete:
      description: 清除当前激活的主题包，恢复为默认主题
      produces:
      - application/json
      responses:
        "200":
          description: 已恢复默认主题
          schema:
            $ref: '#/definitions/models.SuccessResponse'
      security:
      - BearerAuth: []
      summary: 恢复默认主题
      tags:
      - 主题包
    get:
      description: 返回当前激活的主题包详情，无激活主题时返回 null
      produces:
      - application/json
      responses:
        "200":
          description: 激活主题包（可为 null）
          schema:
            $ref: '#/definitions/services.ThemePackResponse'
      security:
      - BearerAuth: []
      summary: 获取激活主题包
      tags:
      - 主题包
    put:
      consumes:
      - application/json
      description: 将指定主题包设为当前使用的主题
      parameters:
      - description: 激活请求
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 设置成功
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: 请求格式错误
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: 主题包不存在
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 设置激活主题包
      tags:
      - 主题包
  /theme-packs/catalog/install:
    post:
      consumes:
      - application/json
      description: 下载远程主题包 JSON，验证 SHA-256 后导入
      parameters:
      - description: 安装请求
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 安装成功
          schema:
            $ref: '#/definitions/services.ThemePackResponse'
        "400":
          description: 校验失败
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 从在线目录安装主题包
      tags:
      - 主题包
  /theme-packs/catalog/refresh:
    post:
      consumes:
      - application/json
      description: 从远程 URL 获取主题目录索引，并标注各主题的安装状态
      parameters:
      - description: 目录请求
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 目录响应
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 获取在线主题目录
      tags:
      - 主题包
...
securityDefinitions:
  BearerAuth:
    description: 输入 "Bearer {token}" 进行认证
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"