# internal/handlers/user_preferences_setting.go
package handlers

import (
	"encoding/json"
	"fmt"
	"net/http"
)

const userPreferencesKey = "user_preferences"

// userPreferencesSetting 用户偏好设置（跨设备同步）
type userPreferencesSetting struct {
	ThemeMode         string  `json:"theme_mode"`
	PlayMode          string  `json:"play_mode"`
	PlaylistViewMode  string  `json:"playlist_view_mode"`
	AudioQuality      string  `json:"audio_quality"`
	LocalCacheMaxSize int64   `json:"local_cache_max_size"`
	Volume            float64 `json:"volume"`
}

var defaultUserPreferences = userPreferencesSetting{
	ThemeMode:         "system",
	PlayMode:          "order",
	PlaylistViewMode:  "grid",
	AudioQuality:      "original",
	LocalCacheMaxSize: 1073741824,
	Volume:            50.0,
}

// GetUserPreferencesSetting 获取用户偏好设置
// @Summary 获取用户偏好设置
// @Description 获取用户跨设备同步的偏好设置，包括主题、播放模式、视图模式、音质、缓存上限和音量。未配置时返回默认值。
// @Tags 设置
// @Produce json
// @Success 200 {object} userPreferencesSetting "用户偏好设置"
// @Security BearerAuth
// @Router /settings/user-preferences [get]
func (h *ConfigHandler) GetUserPreferencesSetting(w http.ResponseWriter, r *http.Request) {
	var cfg userPreferencesSetting
	if err := h.configService.GetJSON(userPreferencesKey, &cfg); err != nil {
		respondJSON(w, http.StatusOK, defaultUserPreferences)
		return
	}
	respondJSON(w, http.StatusOK, cfg)
}

// validateUserPreferences 校验用户偏好设置字段的合法性。
func validateUserPreferences(p *userPreferencesSetting) error {
	validThemeModes := map[string]bool{"system": true, "light": true, "dark": true}
	if p.ThemeMode != "" && !validThemeModes[p.ThemeMode] {
		return fmt.Errorf("theme_mode 必须是 system/light/dark 之一，当前值: %s", p.ThemeMode)
	}

	validPlayModes := map[string]bool{"order": true, "loop": true, "single": true, "random": true, "singlePlay": true}
	if p.PlayMode != "" && !validPlayModes[p.PlayMode] {
		return fmt.Errorf("play_mode 必须是 order/loop/single/random/singlePlay 之一，当前值: %s", p.PlayMode)
	}

	validViewModes := map[string]bool{"grid": true, "list": true}
	if p.PlaylistViewMode != "" && !validViewModes[p.PlaylistViewMode] {
		return fmt.Errorf("playlist_view_mode 必须是 grid/list 之一，当前值: %s", p.PlaylistViewMode)
	}

	validQualities := map[string]bool{"original": true, "128": true, "192": true, "320": true}
	if p.AudioQuality != "" && !validQualities[p.AudioQuality] {
		return fmt.Errorf("audio_quality 必须是 original/128/192/320 之一，当前值: %s", p.AudioQuality)
	}

	if p.Volume < 0 || p.Volume > 100 {
		return fmt.Errorf("volume 必须在 0-100 之间，当前值: %g", p.Volume)
	}

	if p.LocalCacheMaxSize < 0 {
		return fmt.Errorf("local_cache_max_size 不能为负数，当前值: %d", p.LocalCacheMaxSize)
	}

	return nil
}

// UpdateUserPreferencesSetting 保存用户偏好设置
// @Summary 保存用户偏好设置
// @Description 保存用户跨设备同步的偏好设置。客户端登录后拉取、修改偏好时推送，实现多设备间偏好同步。
// @Tags 设置
// @Accept json
// @Produce json
// @Param request body userPreferencesSetting true "用户偏好设置"
// @Success 200 {object} userPreferencesSetting "保存后的用户偏好设置"
// @Failure 400 {object} models.ErrorResponse "请求格式错误"
// @Failure 500 {object} models.ErrorResponse "保存配置失败"
// @Security BearerAuth
// @Router /settings/user-preferences [put]
func (h *ConfigHandler) UpdateUserPreferencesSetting(w http.ResponseWriter, r *http.Request) {
	var req userPreferencesSetting
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		respondError(w, http.StatusBadRequest, "请求格式错误", err)
		return
	}
	if err := validateUserPreferences(&req); err != nil {
		respondError(w, http.StatusBadRequest, err.Error(), nil)
		return
	}
	if err := h.configService.SetJSON(userPreferencesKey, req); err != nil {
		respondError(w, http.StatusInternalServerError, "保存配置失败", err)
		return
	}
	respondJSON(w, http.StatusOK, req)
}